rename traineval -> train
